test/python_tests/requirements.txt
Cython==3.0.2
greenlet==2.0.2
lxml==4.9.3
playwright==1.37.0
psycopg2-binary==2.9.7
pyee==9.0.4
python-docx==0.8.11
python-dotenv==1.0.0
typing_extensions==4.7.1
